<refactor> move Auth hook under hook directory
