+++ front/src/components/comment/InputReply.tsx
@@ -1,4 +1,4 @@
-import { useUserAccountWithSSR } from '@/atoms/userAccount';
+import { useUserAccountWithSSR } from "@/hooks/userAccount";
 import { CommentType } from '@/atoms/comment';
 import React , { useState, useOptimistic } from 'react';
 import { Button,Flex, Input, Text, Textarea} from '@chakra-ui/react';
